requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
pgvector
google-generativeai
gitpython
langchain-text-splitters
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pydantic-settings